+++ hail_search/requirements-test.txt
@@ -1,0 +1,48 @@
+#
+# This file is autogenerated by pip-compile with python 3.10
+# To update, run:
+#
+#    pip-compile hail_search/requirements-test.in
+#
+aiohttp==3.8.4
+    # via pytest-aiohttp
+aiosignal==1.3.1
+    # via aiohttp
+async-timeout==4.0.2
+    # via aiohttp
+attrs==23.1.0
+    # via aiohttp
+charset-normalizer==3.2.0
+    # via aiohttp
+coverage==5.1
+    # via -r hail_search/requirements-test.in
+exceptiongroup==1.1.2
+    # via pytest
+frozenlist==1.3.3
+    # via
+    #   aiohttp
+    #   aiosignal
+idna==3.4
+    # via yarl
+iniconfig==2.0.0
+    # via pytest
+multidict==6.0.4
+    # via
+    #   aiohttp
+    #   yarl
+packaging==23.1
+    # via pytest
+pluggy==1.2.0
+    # via pytest
+pytest==7.4.0
+    # via
+    #   pytest-aiohttp
+    #   pytest-asyncio
+pytest-aiohttp==1.0.4
+    # via -r hail_search/requirements-test.in
+pytest-asyncio==0.21.0
+    # via pytest-aiohttp
+tomli==2.0.1
+    # via pytest
+yarl==1.9.2
+    # via aiohttp
